+++ R/Map_Invalsi.R
@@ -28,6 +28,8 @@
 #' If the level is set to \code{"LAU"}, choosing a limited number of regions is recommended.
 #' By default, \code{c(1,3,5:20)}, i.e. all Italian regions except the provinces of Aosta, Trento and Bozen which have data availability issues.
 #' @param plot Character. The type of map to display; either \code{"mapview"} for interactive maps, or \code{"ggplot"} for static maps. \code{"mapview"} by default.
+#'  !!! Warning. As \code{mapview} turns out to have been taken away from CRAN,
+#'  if it has not been previously installed by the user, \code{plot = "ggplot"} will be forced.
 #' @param pal Character. The palette to use if the \code{"mapview"} mode is chose. \code{"viridis"} by default.
 #' @param WLE Logical. Whether the variable to chose should be the average WLE score rather that the percentage of sufficient tests, if both are available. \code{FALSE} by default
 #' @param col_rev Logical. Whether the scale of the colour palette should be reverted or not, if the \code{mapview} mode is chosen. \code{FALSE} by default
@@ -167,6 +169,8 @@ Map_Invalsi <- function(data = NULL, Year = 2023, subj_toplot = "ITA", grade = 8
   if(only_observed == TRUE){
     res <- res %>% dplyr::filter(!is.na(.data$X))
   }
+
+  if(!"mapview" %in% rownames(utils::installed.packages())) plot <- "ggplot"
 
   if (plot =="mapview"){